feat: Extract Codec & friends from Util into separate modules (#5727)
This fixes #5726 by removing ObjectCodec and ObjectDecoder, their
implementations, as well as CodecCache and it's exception from the Util
module and into their own separate packages, codec-base, codec-builtin
and codec-cache.

(cherry picked from commit 4fdad152ba3a1b37eb45e657d59d15c8f73ae45c)
