+++ codec/builtin/src/main/java/io/deephaven/util/codec/StringDoubleMapCodec.java
