+++ codec/cache/build.gradle
@@ -1,0 +1,12 @@
+plugins {
+    id 'java-library'
+    id 'io.deephaven.project.register'
+}
+
+description 'Codec Cache: Deephaven Codec Cache'
+
+dependencies {
+    api project(":codec-api")
+
+    compileOnly depAnnotations
+}
