+++ codec/cache/gradle.properties
@@ -1,0 +1,1 @@
+io.deephaven.project.ProjectType=JAVA_PUBLIC
