+++ codec/cache/src/main/java/io/deephaven/util/codec/CodecCacheException.java
